requirements.txt
langchain-community>=0.0.10
langchain>=0.1.0
langchain-huggingface>=0.0.5
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
streamlit>=1.28.0
groq>=0.4.0
pydantic>=2.0.0 